requirements.txt
matplotlib==3.9.2
Pillow==11.0.0
pydantic==2.9.2
skimage==0.0
torch==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.4
lpips==0.1.4
scipy==1.11.4
seaborn==0.13.2
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2
autotrain_allmodels.py